+++ server.js
@@ -1,36 +1,20 @@
 var express = require('express'),
-  app = express(),
-  port = process.env.PORT || 8080,
-
-  bodyParser = require('body-parser');
+    app = express(),
+    port = process.env.PORT || 8080,
+    bodyParser = require('body-parser');
 var cors = require('cors');
 var mysql = require('mysql');
-
-
-// Connect to the database
-console.log('Connecting to the database.....');
-var connection = mysql.createConnection({
- database: 'mss_local',  // TODO: change your database name
- host: "localhost",
- user: "root",
- password: "abcd1234!"
-});
-
-connection.connect(function(err) {
- if (err) throw err;
- console.log("Connected to the database!!");
-});
 
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(bodyParser.json());
 app.use(cors());
 
+//importing patient routes
+var patientRoutes = require('./api/routes/patientRoutes'); 
+patientRoutes(app); //register the route
 
-var routes = require('./api/routes/patientRoutes'); //importing route
-routes(app, connection); //register the route
-
+// TODO: importing other routes
+// ...
 
 app.listen(port);
-
-
 console.log('RESTful API server started on: ' + port);
